Improved logging of loading process. Fixed a parsing error for stl binary. Added error message to tree and list tooltips. Added button to open logs folder at config window.

+++ STL_Showcase/Logic/Rendering/Mesh3D.cs
@@ -74,7 +74,7 @@ namespace STL_Showcase.Logic.Rendering
 
             float minX = float.MaxValue, minY = float.MaxValue, minZ = float.MaxValue, maxX = float.MinValue, maxY = float.MinValue, maxZ = float.MinValue;
 
-            int triangleSkip = (TriangleCount > _RepresentativeSalmpleTriangles ? TriangleCount / _RepresentativeSalmpleTriangles : 0);
+            int triangleSkip = (int)(TriangleCount > _RepresentativeSalmpleTriangles ? TriangleCount / _RepresentativeSalmpleTriangles : 0);
             
             for (int i = 0; i < Triangles.Length; i++)
             {
